[Filestore] introducing memory controller for directory handles cache

+++ cloud/filestore/libs/vfs_fuse/fs_impl_list.cpp
@@ -222,7 +222,7 @@ void TFileSystem::ReadDir(
 
                 auto handleChunk = handle->UpdateContent(
                     size,
-                    offset,
+                    static_cast<size_t>(offset),
                     builder.Finish(),
                     version,
                     response.GetCookie());
